requirements.txt
python-dotenv==1.0.0
vaderSentiment==3.3.2
pycoingecko==3.1.0
python-telegram-bot==13.15
feedparser==6.0.10
requests==2.31.0